requirements.txt
Flask
flask-cors
python-dotenv
gunicorn
supabase
PyJWT
requests
groq
runwayml
